First demonstration of a possible overloaded connector/hardware method call from logic. The only downside is that hardware interface methods must avoid the reserved keyword argument "interface".
