+++ config/test.cfg
@@ -25,6 +25,12 @@ logic:
         module.Class: 'jupyterkernel.kernellogic.QudiKernelLogic'
         remoteaccess: True
 
+    testlogic:
+        module.Class: 'test_logic.TestLogic'
+        connect:
+            firsthardware: 'myhardwaremodule'
+            secondhardware: 'myhardwaremodule'
+
 gui:
     tray:
         module.Class: 'trayicon.TrayIcon'
